src/sim/sg13g2/common/requirements.txt
pytest
cocotb
cocotbext-spi
numpy
scipy